pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "konfig"
dynamic = ["version"]
description = "Settings management, pluggable secrets, and run-scoped logging for Python applications."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Matthew Westwood-Hill" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
    "tomli>=2.0; python_version < '3.11'",
    "keyring>=25.0",
    "cryptography>=42.0",
]

[project.optional-dependencies]
aws = ["boto3>=1.34"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
    "isort>=5.13.0",
]

[tool.hatch.version]
path = "src/konfig/_version.txt"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["src/konfig"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.black]
target-version = ["py310"]

[tool.isort]
profile = "black"
